Negative value for zclFrame.payload().at() on x86 (#6702)
* Cast zclFrame.at() to quint8.

See #6670.

* Use cast instead of intermediate variable.

+++ device_js/device_js_wrappers.cpp
@@ -263,7 +263,7 @@ int JsZclFrame::at(int i) const
 {
     if (zclFrame && i >= 0 && i < zclFrame->payload().size())
     {
-        return zclFrame->payload().at(i);
+        return (uint8_t) zclFrame->payload().at(i);
     }
 
     return 0;
